\name{EuclStra-package}
\alias{EuclStra-package}
\alias{EuclStra}
\docType{package}
\title{
\packageTitle{EuclStra}
}
\description{
\packageDescription{EuclStra}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EuclStra}
\packageIndices{EuclStra}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{EuclStra}

Maintainer: \packageMaintainer{EuclStra}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
